:truck: move `DataWidget` into its own module
